+++ SmartFridgeApp.API/Recipes/FindRecipes/FindRecipesRequest.cs
@@ -4,6 +4,6 @@ namespace SmartFridgeApp.API.Recipes.FindRecipes
 {
     public class FindRecipesRequest
     {
-        public List<int> FoodProducts { get; set; }
+        public List<short> FoodProducts { get; set; }
     }
 }
